requirements.txt
pip~=22.0.4
attrs~=23.2.0
wheel~=0.37.1
cryptography~=42.0.5
Pillow~=9.2.0
jinja2~=3.1.4
idna~=3.3
multidict~=6.0.5
yarl~=1.9.4
aiohttp~=3.9.5
aiosignal~=1.3.1
frozenlist~=1.4.1
pytest~=8.0.2
discord~=2.3.2
requests~=2.28.1
python-dotenv~=1.0.1
polars~=1.21.0
